refactor(mitm-addon): extract runner flush lifecycle owner
