+++ crates/runner/mitm-addon/tests/conftest.py
@@ -32,6 +32,7 @@ import logging_utils
 import mitm_addon
 import platform_api
 import registry
+import runner_flush_lifecycle
 import upstream_admission
 import upstream_destination_binding
 import usage
@@ -57,7 +58,7 @@ def _reset_module_state() -> Iterator[None]:
     registry.reset_cache_for_tests()
     upstream_destination_binding.reset_for_tests()
     upstream_admission.reset_upstream_destination_resolution_cache_for_tests()
-    mitm_addon.reset_runner_usage_flush_state_for_tests()
+    runner_flush_lifecycle.reset_runner_usage_flush_state_for_tests()
     upstream_admission.reset_tls_admission_state_for_tests()
     platform_api.configure_client_headers(client_session_id="", client_version="")
     clear_auth_state()
@@ -67,7 +68,7 @@ def _reset_module_state() -> Iterator[None]:
     usage.reset_usage_buffer_for_tests()
     logging_utils.reset_log_writer_for_tests()
     yield
-    mitm_addon.reset_runner_usage_flush_state_for_tests()
+    runner_flush_lifecycle.reset_runner_usage_flush_state_for_tests()
     usage.reset_usage_buffer_for_tests()
     logging_utils.reset_log_writer_for_tests()
     auth_base_forwarder.reset_forward_request_state_for_tests()
