+++ crates/runner/src/proxy/mod.rs
@@ -36,7 +36,9 @@
 //! for `usage-pending`, then calls `stop`.
 //!
 //! Addon-side details live in `crates/runner/mitm-addon/src/mitm_addon.py`
-//! (SIGUSR1 worker and request handling),
+//! (mitmproxy hook orchestration),
+//! `crates/runner/mitm-addon/src/runner_flush_lifecycle.py` (SIGUSR1 worker and
+//! request handling),
 //! `crates/runner/mitm-addon/src/usage/counters.py` (`usage-pending`),
 //! `crates/runner/mitm-addon/src/registry.py` (registry loading), and
 //! `crates/runner/mitm-addon/src/jsonl_writer.py` (accepted-write flush
